New export helper
A bunch of the specs that exercise the `Export` classes use inline
fixtures. Here we consolidate the helper methods that dirve those into
one place.
