+++ spec/models/export/activity_actuals_columns_spec.rb
@@ -12,7 +12,7 @@ RSpec.describe Export::ActivityActualsColumns do
     @q3_report = create(:report, financial_quarter: 3, financial_year: 2020)
     @q4_report = create(:report, financial_quarter: 4, financial_year: 2020)
 
-    create_fixtures(
+    actuals_from_table(
       <<~TABLE
         |transaction|report|financial_period|value|
         | Actual    |q1    | q1             |  100|
@@ -248,37 +248,5 @@ RSpec.describe Export::ActivityActualsColumns do
         .to have_received(:new)
         .exactly(@activities.count).times
     end
-  end
-
-  def value_for_header(header_name)
-    values = subject.rows.fetch(@activity.id)
-    values[subject.headers.index(header_name)]
-  end
-
-  def create_fixtures(table)
-    CSV.parse(table, col_sep: "|", headers: true).each do |row|
-      case row["transaction"].strip
-      when "Actual"
-        create(:actual, fixture_attrs(row))
-      when "Adj. Act."
-        create(:adjustment, :actual, fixture_attrs(row))
-      when "Adj. Ref."
-        create(:adjustment, :refund, fixture_attrs(row))
-      when "Refund"
-        create(:refund, fixture_attrs(row))
-      else
-        raise "don't know what to do"
-      end
-    end
-  end
-
-  def fixture_attrs(row)
-    {
-      parent_activity: @activity,
-      value: row["value"].strip,
-      financial_quarter: row["financial_period"][/\d/],
-      financial_year: 2020,
-      report: instance_variable_get("@#{row["report"].strip}_report"),
-    }
   end
 end
