+++ spec/models/export/activity_variance_column_spec.rb
@@ -68,40 +68,4 @@ RSpec.describe Export::ActivityVarianceColumn do
       expect(variance_for_activity).to eq BigDecimal(20000 - 40000)
     end
   end
-
-  private
-
-  def actuals_from_table(table)
-    CSV.parse(table, col_sep: "|", headers: true).each do |row|
-      case row["transaction"].strip
-      when "Actual"
-        create(:actual, fixture_attrs(row))
-      when "Refund"
-        create(:refund, fixture_attrs(row))
-      else
-        raise "don't know what to do"
-      end
-    end
-  end
-
-  def forecasts_for_report_from_table(report, table)
-    CSV.parse(table, col_sep: "|", headers: true).each do |row|
-      ForecastHistory.new(
-        @activity,
-        report: report,
-        financial_quarter: row["financial_quarter"].to_i,
-        financial_year: row["financial_year"].to_i,
-      ).set_value(row["value"].to_i)
-    end
-  end
-
-  def fixture_attrs(row)
-    {
-      parent_activity: @activity,
-      value: row["value"].strip,
-      financial_quarter: row["financial_period"][/\d/],
-      financial_year: 2021,
-      report: instance_variable_get("@#{row["report"].strip}_report"),
-    }
-  end
 end
