+++ spec/rails_helper.rb
@@ -49,6 +49,7 @@ RSpec.configure do |config|
   config.include StripAttributes::Matchers
   config.include TimeTravelHelpers
   config.include TransactionHelpers
+  config.include ExportHelpers
 
   # Remove this line if you're not using ActiveRecord or ActiveRecord fixtures
   config.fixture_path = "#{::Rails.root}/spec/fixtures"
